Removed binary compilation out of main, and into ILanguageBackend.
