+++ src/ShaderGen/Glsl/Glsl330Backend.cs
@@ -6,7 +6,14 @@ namespace ShaderGen.Glsl
     public class Glsl330Backend : GlslBackendBase
     {
         public override string GeneratedFileExtension => "330.glsl";
-        
+        public override bool CompilationToolsAreAvailable() {
+            return false;
+        }
+
+        public override bool CompileCode(string shaderPath, string entryPoint, ShaderFunctionType type, out string path) {
+            throw new System.NotImplementedException();
+        }
+
         public Glsl330Backend(Compilation compilation) : base(compilation)
         {
         }
