+++ src/ShaderGen/ILanguageBackend.cs
@@ -1,6 +1,9 @@
 ﻿namespace ShaderGen {
     public interface ILanguageBackend {
         string GeneratedFileExtension { get; }
+
+        bool CompilationToolsAreAvailable();
+        bool CompileCode(string shaderPath, string entryPoint, ShaderFunctionType type, out string path);
         
         void InitContext(string setName);
         void AddStructure(string setName, StructureDefinition sd);
