+++ src/ShaderGen/LanguageBackend.cs
@@ -9,7 +9,9 @@ namespace ShaderGen
     public abstract class LanguageBackend : ILanguageBackend
     {
         public abstract string GeneratedFileExtension { get; }
-        
+        public abstract bool CompilationToolsAreAvailable();
+        public abstract bool CompileCode(string shaderPath, string entryPoint, ShaderFunctionType type, out string path);
+
         protected readonly Compilation Compilation;
 
         internal class BackendContext
